Add configurable residual GNN options and ablation script
